Fix 2039 warnings The remaining warnings are all in external libraries

+++ libs/threadpool/src/queue.cc
@@ -51,7 +51,7 @@ void QueueImpl::jobFinished()
 int QueueImpl::getCount()
 {
   boost::mutex::scoped_lock lock(mut);
-  return count;
+  return static_cast<int>(count);
 }
 
 ////////////////////////////////////////////////////////////////////////
